+++ Snowflake.Data/Core/Session/SFSessionProperty.cs
@@ -152,6 +152,8 @@ namespace Snowflake.Data.Core
         MINTLS,
         [SFSessionPropertyAttr(required = false, defaultValue = "tls13")]
         MAXTLS,
+        [SFSessionPropertyAttr(required = false, defaultValue = "false")]
+        HONORSESSIONTIMEZONE,
     }
 
     class SFSessionPropertyAttr : Attribute